Revert "fix/LaggingReplicaReassignmentGoal: store state in clusterModel"
This reverts commit 615d5815d0f510acf9872a84c9a8ea8bde84d972.
