+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/analyzer/goals/LaggingReplicaReassignmentGoal.java
@@ -6,6 +6,7 @@ package com.linkedin.kafka.cruisecontrol.analyzer.goals;
 
 import java.util.Arrays;
 import java.util.ArrayList;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutionException;
 import java.util.HashMap;
 import java.util.List;
@@ -25,7 +26,6 @@ import com.linkedin.kafka.cruisecontrol.exception.OptimizationFailureException;
 import com.linkedin.kafka.cruisecontrol.model.Broker;
 import com.linkedin.kafka.cruisecontrol.model.ClusterModel;
 import com.linkedin.kafka.cruisecontrol.model.ClusterModelStats;
-import com.linkedin.kafka.cruisecontrol.model.PartitionInfoWrapper;
 import com.linkedin.kafka.cruisecontrol.monitor.ModelCompletenessRequirements;
 import org.apache.kafka.clients.admin.AdminClient;
 import org.apache.kafka.clients.admin.NewPartitionReassignment;
@@ -52,9 +52,9 @@ public class LaggingReplicaReassignmentGoal extends AbstractGoal {
 
     private List<PartitionInfo> _laggingPartitions;
 
-    protected HashMap<PartitionInfoWrapper, Long> _laggingPartitionsMap;
+    protected ConcurrentHashMap<PartitionInfoWrapper, Long> _laggingPartitionsMap;
 
-    private HashMap<PartitionInfoWrapper, Long> _newLaggingPartitionsMap;
+    private ConcurrentHashMap<PartitionInfoWrapper, Long> _newLaggingPartitionsMap;
 
     private long _maxReplicaLagMs;
 
@@ -64,17 +64,15 @@ public class LaggingReplicaReassignmentGoal extends AbstractGoal {
 
     @Override
     public void configure(Map<String, ?> configs) {
-        LOG.info("Configuring LaggingReplicaReassignmentGoal");
         _parsedConfig = new KafkaCruiseControlConfig(configs, false);
         _adminClient = createAdminClient(parseAdminClientConfigs(_parsedConfig));
         _balancingConstraint = new BalancingConstraint(_parsedConfig);
         _numWindows = _parsedConfig.getInt(MonitorConfig.NUM_PARTITION_METRICS_WINDOWS_CONFIG);
         _minMonitoredPartitionPercentage = _parsedConfig.getDouble(MonitorConfig.MIN_VALID_PARTITION_RATIO_CONFIG);
-        // _laggingPartitionsMap = new HashMap<PartitionInfoWrapper, Long>();
+        _laggingPartitionsMap = new ConcurrentHashMap<PartitionInfoWrapper, Long>();
         _maxReplicaLagMs = (long) configs.get(AnalyzerConfig.MAX_LAGGING_REPLICA_REASSIGN_MS);
         _laggingPartitions = new ArrayList<PartitionInfo>();
         _laggingRecoveryNeeded = false;
-        LOG.info("Configured LaggingReplicaReassignmentGoal");
     }
 
     @Override
@@ -156,22 +154,26 @@ public class LaggingReplicaReassignmentGoal extends AbstractGoal {
 
     void checkIfReplicasLagging(ClusterModel clusterModel) throws OptimizationFailureException {
         long currentTimeMillis = System.currentTimeMillis();
-        _newLaggingPartitionsMap = new HashMap<PartitionInfoWrapper, Long>();
+        _newLaggingPartitionsMap = new ConcurrentHashMap<PartitionInfoWrapper, Long>();
         LOG.info("Checking for lagging replicas");
-        LOG.info("Admin client details {}, {}", _adminClient.toString(), _adminClient.hashCode());
-        _laggingPartitionsMap = clusterModel.getLaggingPartitionsMap();
+        if (_laggingPartitionsMap == null) {
+            _laggingPartitionsMap = new ConcurrentHashMap<PartitionInfoWrapper, Long>();
+        }
+        //List<PartitionInfo> laggingPartitionInfos = clusterModel.getPartitionsWithLaggingReplicas();
+        //_laggingPartitionsMap.entrySet().removeIf(e -> !laggingPartitionInfos.contains(e.getKey()._pi));
         for (PartitionInfo partition: clusterModel.getPartitionsWithLaggingReplicas()) {
             LOG.info(partition.toString());
             PartitionInfoWrapper piw = new PartitionInfoWrapper(partition);
             long lastSeenTime = _laggingPartitionsMap.getOrDefault(piw, currentTimeMillis);
             if (currentTimeMillis - lastSeenTime >= _maxReplicaLagMs) {
-                LOG.info("Partition {} has been lagging for past {} minutes", partition.toString(), (currentTimeMillis - lastSeenTime) / (60 * 1000));
+                LOG.info("Partition {} has been lagging for past {} minutes", partition.toString(), 
+                            (currentTimeMillis - lastSeenTime) / (60 * 1000));
                 _laggingRecoveryNeeded = true;
                 _laggingPartitions.add(partition);
             }
             _newLaggingPartitionsMap.put(piw, lastSeenTime);
         }
-        clusterModel.setLaggingPartitionsMap(_newLaggingPartitionsMap);
+        _laggingPartitionsMap = _newLaggingPartitionsMap;
         LOG.info("Lagging partitions map: {}  on thread after {}", _laggingPartitionsMap.toString(), Thread.currentThread().getName());
     }
 
@@ -202,6 +204,55 @@ public class LaggingReplicaReassignmentGoal extends AbstractGoal {
             } catch (InterruptedException | ExecutionException e) {
                 LOG.error("Unable to move replicas onto same brokers");
             }
-        }    
+        }
+        
     }
+    protected static class PartitionInfoWrapper {
+
+        PartitionInfo _pi;
+
+        public PartitionInfoWrapper(PartitionInfo pi) {
+            this._pi = pi;
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (o == this) {
+                return true;
+            }
+            if (!(o instanceof PartitionInfoWrapper)) {
+                return false;
+            }
+            PartitionInfoWrapper partitionInfoWrapperObj = (PartitionInfoWrapper) o;
+            PartitionInfo p2 = partitionInfoWrapperObj._pi;
+            if (_pi.topic().equals(p2.topic()) && _pi.partition() == p2.partition() 
+                && _pi.leader().id() == p2.leader().id() && _pi.inSyncReplicas().length == p2.inSyncReplicas().length) {
+                Set<Integer> p2ISRSet = Arrays.stream(p2.inSyncReplicas()).map(isr -> isr.id()).collect(Collectors.toSet());
+                if (Arrays.stream(_pi.inSyncReplicas()).allMatch(isr -> p2ISRSet.contains(isr.id()))) {
+                    return true;
+                }
+            }
+            return false;
+        }
+
+        @Override
+        public int hashCode() {
+            final int prime = 31;
+            int result = 1;
+            result = prime * result + ((_pi.topic() == null) ? 0 : _pi.topic().hashCode());
+            result = prime * result + _pi.partition();
+            result = prime * result + _pi.leader().id();
+            for (Node n: _pi.inSyncReplicas()) {
+                result = prime * result + n.id();
+            }
+            return result;
+        }
+
+        @Override
+        public String toString() {
+            return _pi.toString();
+        }
+
+    }
+    
 }
